refactored MoveStack
